Afisha events serializer is shared now and stored in serializer.py
