--- apps/afisha/serializers/__init__.py
@@ -1,3 +1,0 @@
-from .event import EventFestivalSerializer, EventRegularSerializer
-
-__all__ = (EventRegularSerializer, EventFestivalSerializer)
